+++ CHANGELOG.md
@@ -5,8 +5,8 @@
 * Adds new resource `v1/config/LogScaleAction`
 * Adds new resource `v1/config/LogScaleAlert`
 * Adds new field `pool.allocation.fixed_values` to `v1/config/ResourcePools`
-* New unstable state: metric-usages-by-metric-name
-* New unstable state: metric-usages-by-label-name
+* Adds new state command `metric-usages-by-metric-name`
+* Adds new state command `metric-usages-by-label-name`
 
 ## v1.9.0